refactored with utilty and added aria label for months
