+++ src/containers/Calendar.tsx
@@ -6,6 +6,7 @@ import styles from "../styles/stylesCalendar.css";
 import MonthPicker from "../components/MonthPicker";
 import DaysHeading from "../components/DaysHeading";
 import DatesOfMonth from "../components/DatesOfMonth";
+import { keyDownHandler } from "../utility/functions"
 
 const Calendar: React.FC = () => {
   const [clickedDate, setClickedDate] = useState({});
@@ -24,71 +25,13 @@ const Calendar: React.FC = () => {
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [dateObject]);
 
-  console.log("Selected date is: ", clickedDate);
-
-  const keyDownHandler = (event: React.KeyboardEvent) => {
-    event.preventDefault();
-    const eventTarget = event.target as HTMLElement
-    const currentElementId = eventTarget.id.slice(-2) as string;
-
-    const sliceChecker = (elementId: string): number => {
-      if (elementId[0] === "-") {
-        return +elementId.slice(-1)
-      } else {
-        return +elementId
-      }
-    }
-
-    if (event.key === "ArrowLeft") {
-      if ((document.getElementById(`button-${sliceChecker(currentElementId) - 1}`) as HTMLButtonElement) === null) {
-        return;
-      } else {
-        (document.getElementById(`button-${sliceChecker(currentElementId) - 1}`) as HTMLButtonElement).focus();
-      }
-    }
-    else if (event.key === "ArrowRight") {
-      if ((document.getElementById(`button-${sliceChecker(currentElementId) + 1}`) as HTMLButtonElement) === null) {
-        return;
-      } else {
-        (document.getElementById(`button-${sliceChecker(currentElementId) + 1}`) as HTMLButtonElement).focus();
-      }
-    }
-    else if (event.key === "ArrowDown") {
-      if ((document.getElementById(`button-${sliceChecker(currentElementId) + 7}`) as HTMLButtonElement) === null) {
-        if (eventTarget.parentElement?.parentElement?.nextElementSibling?.firstElementChild?.firstElementChild as HTMLButtonElement) {
-          (eventTarget.parentElement?.parentElement?.nextElementSibling?.firstElementChild?.firstElementChild as HTMLButtonElement).focus();
-        } else {
-          return;
-        }
-      } else {
-        (document.getElementById(`button-${sliceChecker(currentElementId) + 7}`) as HTMLButtonElement).focus();
-      }
-    }
-    else if (event.key === "ArrowUp") {
-      if ((document.getElementById(`button-${sliceChecker(currentElementId) - 7}`) as HTMLButtonElement) === null) {
-        if (eventTarget.parentElement?.parentElement?.previousElementSibling as HTMLButtonElement) {
-          (document.getElementById("button-1") as HTMLButtonElement).focus();
-        }
-      } else {
-        (document.getElementById(`button-${sliceChecker(currentElementId) - 7}`) as HTMLButtonElement).focus();
-      }
-    }
-    else if (event.key === "Enter" || event.key === "Spacebar" || event.key === " ") {
-      eventTarget.click();
-    } else if (event.key === "Home") {
-      (document.getElementById("button-1") as HTMLButtonElement).focus();
-    } else if (event.key === "End") {
-      (document.getElementById(`button-${dateObject.dates}`) as HTMLButtonElement).focus();
-    }
-}
-
   return (
     <div className={styles.calendarContainer} role="application">
       <MonthPicker
         month={dateObject.month}
         year={dateObject.year}
         setDateObject={setDateObject} />
-      <table id='calendar-table' className={styles.calendarTableContainer} role="presentation" onKeyDown={keyDownHandler}>
+      <table id='calendar-table' className={styles.calendarTableContainer} role="presentation" onKeyDown={(event) => keyDownHandler(event, dateObject.dates)}>
         <DaysHeading />
         <DatesOfMonth
           year={dateObject.year}
